+++ src/main/java/wooteco/subway/domain/path/Shortest.java
@@ -1,4 +1,6 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
+
+import wooteco.subway.domain.Station;
 
 public interface Shortest {
     Path getShortestPath(final Station source, final Station target, final int fare, final int age);
