+++ src/main/java/wooteco/subway/domain/path/ShortestPath.java
@@ -1,4 +1,6 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
+
+import wooteco.subway.domain.Station;
 
 public interface ShortestPath {
     Path getPath(final Station source, final Station target, final int fare, int age);
