+++ src/test/java/wooteco/subway/domain/PathTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.domain.path.ShortestPath;
+import wooteco.subway.domain.path.SubwayGraph;
 
 import java.util.List;
